agregado metodo delete y optimizados todos los controladores de gruplac y cvlac

+++ cvlac/gruplac_models/DBmodel.py
@@ -1,5 +1,6 @@
 
 from cvlac.db_gruplac import Base, engine
+from sqlalchemy_utils import database_exists, create_database
 
 from cvlac.gruplac_models.Basico import Basico
 from cvlac.gruplac_models.CursoDoctorado import CursoDoctorado
@@ -13,7 +14,9 @@ from cvlac.gruplac_models.ProgramaMaestria import ProgramaMaestria
 from cvlac.gruplac_models.MetaDB import MetaDB
 
 def create_db():
-    #######################
-    Base.metadata.drop_all(engine)
-    Base.metadata.create_all(engine)
-    ###########################
+    if not database_exists(engine.url):
+        create_database(engine.url)
+        Base.metadata.create_all(engine)
+    else:
+        Base.metadata.drop_all(engine)
+        Base.metadata.create_all(engine)
